fix: use u64 (instead of i64)
Fields include: limit, blocks_limit, snarks_limit, trans_limit, epoch, current_epoch, slot_in_epoch, selected_epoch, height, block_height, nonce, and slot

Fixes Granola-Team/mina-block-explorer#600

Related:
- graphql-rust/graphql-client#266
- graphql-rust/graphql-parser#53

+++ src/summary/models.rs
@@ -5,8 +5,8 @@ use serde::{Deserialize, Serialize};
 pub struct BlockchainSummary {
     pub blockchain_length: u64,
     pub circulating_supply: String,
-    pub epoch: u32,
-    pub slot: u32,
+    pub epoch: u64,
+    pub slot: u64,
     pub global_slot: i64,
     pub staking_epoch_ledger_hash: String,
     pub total_currency: String,
